+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * raised minimum required PHP version to 7.0
   * dropped support for SVN repositories, only Git repositories are supported by now
   * implemented #10 replace herrera-io/version as it is not maintained any more
+  * fixed #7 Git repository does not respect current branch
 
 
 1.0.0 (2013-11-05)